I now have working tests around determiners and complementizers.
Also, I changed the semantics of the indefinite article so that it does not create new subclasses. This function is now exclusive to the keyword 'is'. The motivation for this change as to allow named but untyped nouns to be used as parameters.

The pragmatics module will eventually hold libraries and has been restructured to fit that purpose.